fix panic in tests when server integration tests are not enabled

+++ integration_tests/ssh3_test.go
@@ -47,7 +47,9 @@ var _ = BeforeSuite(func() {
 
 var _ = AfterSuite(func() {
 	CleanupBuildArtifacts()
-	serverSession.Terminate()
+	if serverSession != nil {
+		serverSession.Terminate()
+	}
 })
 
 var _ = Describe("Testing the ssh3 cli", func() {
